mlpipe/stacks/aimodels/landfill/detr50/inference/code/requirements.txt
transformers
pytorch-lightning
pillow
timm